Modularize DB caching components
